Remove a bit of code duplication related to refinement of variables and properties
Summary: We have a bit of duplication when we do refinement of variable and  instance and class variables, and as a consequence the duplicated pieces of code started to diverge slightly (and in particular, refinement due to null checks work for let-variables and `$$` but say `is_int`, `is_string` etc. checks don't).  This diff attempts to remove the duplication.

Reviewed By: andrewjkennedy

Differential Revision: D8009675

fbshipit-source-id: 0148de58dd95ce980a301ffdb6b84fd72ab3e5a0