Add TypeStructure::toStringforDisplay for as-expressions
Summary:
Error messages in `as` expressions should reflect the types that users write in Hack code. In this diff, we add `TypeStructure::toStringForDisplay`, which emits the types as a user would expect to see them (primarily stripping the `HH\` prefix).

We follow the same rules as in `type-constraint.cpp`[1], namely, we leave hack arrays with the prefix (I don't know the historical reasons behind that).

Reviewed By: oulgen

Differential Revision: D7962018

fbshipit-source-id: 9cfe0f2f65b83bef348a3e72e87123f2aae575f9