make sure persistent strings/arrays live in uncounted region
Summary:
When jemalloc 5 is used, we try to lay out the memory so that uncounted heap objects live below a certain address.
When static arrays are not allocated using low_malloc() (either due to the runtime option, or due to low arena being full), this makes it fall back to uncounted_malloc().
Also increase the maximize size for uncounted arena.

When jemalloc 5 is not used, every allocator eventaully calls plain malloc(), and thus no assumption can be made about address ranges.

Reviewed By: markw65

Differential Revision: D8014874

fbshipit-source-id: d3b363f2432ff04afd625fe4b09d73f6400f2357