Handle persistent client commands needing full recheck
Summary:
So far, we've been able to handle everything that was sent our way almost immediately.

But what happens when persistent client client sends a command that we cannot immediately handle (because it needs full recheck)? It's blocked on receiving the result, so it won't even send any other commands before receiving an answer. This is not good.

At some point, we'll need to rewrite `clientLsp.ml` to be able to do it in non-blocking way, and for server to be able to process persistent client commands out of order, but for now I'll solve it differently - by trying to (drastically) reduce the amount of times it can happen.

There are very few commands that actually need full recheck, and persistent client doesn't even execute most of them:

* it doesn't ask for `STATUS` (it depends on notifications instead)
*  it doesn't do `BUILD` (and build is going away, any moment now!)
* `SUBSCRIBE_DIAGNOSTICS` is made to not require full check in D7923668

Only one left is `FIND_REFS` (`textDocument/references`). And this runs so rarely[1] that I don't think it's worth worrying about now.

Instead I put it in a queue (that has always at most one element), and stop processing any further commands until this one is handled. This means that if someone does FIND_REFS in Nuclide during a long typecheck, their experience degrades to current version of “nothing works until recheck completes”, but... meh.

This is very similar to {D7923730}, except it blocks client commands not until current job can be cancelled (milliseconds - seconds), but until it is finished (possibly minutes).

Reviewed By: alexchow

Differential Revision: D7923770

fbshipit-source-id: 59524396d5ddbf1f7ff6913c78948b6cb611a8cd