Do not automatically restart full checks cancelled because of IDE edit commands
Summary:
Here comes the big (?) disappointment of this diff stack: while we can do things during typechecking, the most important (?) and time sensitive sequence of IDE queries (edit + autocomplete) requires us to STOP typechecking (to apply the edit in thread safe way). We can do it, but there is a constant cost to starting and stopping those jobs, which can exceed the time Nuclide allots us to return the results and results in a very flaky experience.

I tried a bunch of heuristics to try to automatically determine when to restart and none of them were particularly satisfying, so instead I opt to transfer this responsibility to the user - when you start editing files you cancel current recheck and it will be only restarted when you save any file again.

Reviewed By: alexchow

Differential Revision: D7944023

fbshipit-source-id: 0af0123f00c1c0b971c55bef7cef8b57c6a1b701