HHVM Debugger: Don't store SourceRootInfo in session object
Summary:
This fixes a bug in the debugger session that was causing HHVM to segfault sometimes when the debugger client detached.

SourceRootInfo allocates an HPHP::String (which is in thread-local request memory). If the worker thread that allocated this exits before the session is deleted, we assert on dbg builds.

Reviewed By: markw65

Differential Revision: D8033702

fbshipit-source-id: 63505e34f84e13259e158423be5b73d7da9d9c0e