Delete non hhvm compat mode error checks
Summary:
There are a bunch of errors that we raise as "extra" errors. These aren't turned on in hh_server or in hhvm codegen, so they kind of are just sitting here wasting time. I've checked and as far as I can tell, all of them are implemented in hh_server already, so we shouldn't need them.

Removing them amounts to a 7% perf win according to the perf tool on 100 small files(running each 1000 times). Full_fidelity_parser_errors goes down from 25% of the time spent to 18%.

Reviewed By: oulgen

Differential Revision: D8014342

fbshipit-source-id: b6d92505cc9613f98819e4d12996a08f258a941f