Emit Hack Array Compat notices in is-expressions
Summary: To migrate away from `is_vec` and `is_dict`, we must preserve the notices emitted when they're passed varray/darray, respectively.  The heavy lifting was done in D7237938; we call in to the same functions and add an extra nullability check afterward if necessary. In HHBBC, we can only optimize away the bytecode if we know the types match; otherwise we might miss a notice.

Reviewed By: viratyosin

Differential Revision: D7994022

fbshipit-source-id: 9367232161b70521bc7af08de3a36faeb295b6a2