Use PGO for object-method calls in more cases
Summary:
If the JIT knew an interface implemented by the object,
it didn't profile the method dispatch (via MethProfile) to try to determine if a
dynamic check could be used to emit a faster call.  This diff enables
MethProfile and associated method-dispatch optimizations to be used in such
cases for object-method calls.

Reviewed By: markw65

Differential Revision: D8022663

fbshipit-source-id: adcb8b448f60644750436a590611a970dc63ad38