Remove extraneous field in static type structure
Summary:
In the static type structure generated for is/as expressions, we were adding an extra "typevars" field unconditionally. This is because
- we added "HH\_" to the list of `tparams` for it to generate the generic kind properly
- the `is_typedef` flag would add the "typevars" field if the `tparams` list was populated.

In this diff, we forego any magic and hardcode the underscore as a type parameter. We also remove the `is_typedef` flag, because these anonymous type structures aren't typedefs.

Reviewed By: oulgen

Differential Revision: D8046819

fbshipit-source-id: 547fb67425928c50ced53d81c460964f30b25d33