Refine types in as-expressions
Summary:
This uses the same logic to refine types in as-expressions as we do for is-expressions. We extract `safely_refine_type` into a module-level function and reuse it when typing as-expressions.

We also add some tests that display the nuance of this behavior, such as with wildcards and refinement of individual elements in tuples.

Reviewed By: oulgen

Differential Revision: D8045669

fbshipit-source-id: 1505e3eccae2958b9dcc6c4dd6682dd88f70becc