Limit the scope of Nonfatal_rpc_exception try...with
Summary:
The intention of `Nonfatal_rpc_exception` is to silently ignore failures during handling of commands (because there is no need to drop everything on the floor if current “go-to-def” request ran into some edge case in parser and thrown an exception).

I don't think we want to do the same thing for failures to `ping` client, or for `sending_responses` to client. There are other try-withs dealing with those.

Reviewed By: jamesjwu

Differential Revision: D8051984

fbshipit-source-id: 6dc940cf6bd0334793339230b1c723c5861a1997