Wrap continuations represeting client commands in try...with
Summary:
This is broken in a major way as of D7923599 and needs to be fixed before next release. Current trunk version has server die in some completely unnecessary situations (like doing Ctrl+C on a client before it received result). Sorry!

I changed `ServerMain.handle_connection` to (possibly) return a continuation instead of handling the command. But there is a very important try-with inside `handle_connection`, that now can be avoided.

I fix it here by wrapping the continuations in same try-withs.

Reviewed By: jamesjwu

Differential Revision: D8052005

fbshipit-source-id: fcb8e90f5c6211afaeb60315bf4f319742246ddc