Explicitly trigger full checks in integration tests
Summary:
There is a pattern in integration tests -  whenever I want the server to perform a full recheck, I ask it for full error list (which immediately forces this recheck).

I want to change the scheduling algorithm so that this will not be guaranteed (the request can be stored in environment to be handled later). I change the tests not to depend on this behavior.

Reviewed By: jamesjwu

Differential Revision: D8052008

fbshipit-source-id: 1fb5ed9b65b6498c1e29b7d1054bbf55f8ffa19f