Resolve type structure statically when possible for is/as irgen even if it is partial resolution
Summary:
This diff adds static type structure resolution to is/as expressions at the irgen phase. Even if we only partially resolve the type structure, we still get a win from it.

Difference from D7902884 (the one that got reverted):

The default value for `persistent` is now set to `true` as if the value of `persistent` provided by the caller of `TypeStructure::resolve`, we can never set `persistent` to be true as we only manipulate is by `&=`. Hence, in this diff, we ignore the user provided value and always assume it is persistent and while we are resolving we can only unset persistent.
Browsed through all the user sites of `resolve` and they all provide `true` anyway, so there should be no difference in behavior.

This diff also takes a different approach from D7986890 where we inline invalid type detection to resolution phase.

Reviewed By: markw65

Differential Revision: D8038394

fbshipit-source-id: cded851d7fdedaea4ebcd623d5a68ca709fc979f