Return kind and position of nodes rather than node content
Summary:
The current output is the JSON-serialized version of the entire node. However, this isn't actually enough information to identify the node in its source file; it could be confused with any identical-content node. (This is especially common for `Missing` nodes.) Additionally, producing the entire node as output is too much output and takes too long on our codebase.

Instead, just return the kind and offset range of the node. The kind is necessary in the event that one node contains another and it happens to have the same range. The offset range is inclusive on both sides.

Reviewed By: pittsw

Differential Revision: D7903885

fbshipit-source-id: 6e32263f5eedd9324e49d3deab3a970b9e7f28f1