Partition blocks by area_idx in layoutBlocks
Summary: Vtext was being passed to layoutBlocks so that pgoLayout could partition the blocks based on the code area pointed to by the block's area_idx. But pgoLayout is only used for optimized translations, for which the 3 code areas are always different. This diff changes the partition to be based on area_idx directly and drops the Vtext parameter.

Reviewed By: paulbiss

Differential Revision: D8023107

fbshipit-source-id: 6dc1a29271421a89225df730e667e022d511bae8