Introduce __ReturnsVoidToRx
Summary: This diff adds support for `__ReturnsVoidToRx` attribute which forces return type of the function to be `void` in reactive context. As a consequence methods annotated with this attribute are allowed to return mutably borrowed values (specifically `$this`) since rx caller will see return type as void and won't be able to use it.

Reviewed By: KendallHopkins

Differential Revision: D8038195

fbshipit-source-id: b1903d08c9f2d8b8e4e73239646d2e6b0851890f