Profile instance bits in method profiler
Summary:
MethProfile may decide to emit dynamic checks to verify that the runtime class
is an instance of a certain class or implements a certain interface.  Those
checks can be optimized to use InstanceOfBitmask if the class/interface is
assigned an instance bit.  This diff bumps the instance-bits profile counters in
those cases in MethProfile.

Reviewed By: markw65

Differential Revision: D8067191

fbshipit-source-id: 37c400b564fd9a1b90b216009caf8948145476b1