Resolve public interface methods and support FCallAwait on an abstract base
Summary:
In www async functions on interfaces and abstract functions on base classes are defined as returning Awaitable but not marked as async as this is considered an implementation detail and these functions lack implementations. In hhbbc when determining if a FuncFamily is eligible for FCallAwait we were previously considering abstract functions although these functions either cannot be called (instance methods) or always fatal (static methods) and therefore need not be considered.

With this fixed we are still unable to issue FCallD or FCallAwait reductions on calls where the base is known to be an object which implements a particular interface. In large part this is because given an interface and an instance it is difficult to tell whether the call was made with the correct context, or if there may be a private method on a parent class whether the call was made to the private version. This is because our usual strategy of comparing the context class with the called class won't work when the called class is only known to implement an interface.

We could add some additional heuristics to allow certain private and protected methods to be considered part of an interface but as interfaces almost always define only public methods (I believe we now allow protected methods on interfaces) this would serve little purpose. To sidestep these issues methods for which a private ancestor may appear on any implementing class, and methods with non-public visibility are excluded.

As a final optimization we observe that while hhbbc will not track intersection types of object classes and interfaces many interfaces and methods are always found together although not declared on the interface itself. To handle this we collect the interface methods by observing which common methods implementers define, rather than by restricting the list of methods to those declared on the interface.

With these changes we see that both calls in the below test program become FCallAwait when compiled in repo mode.

```
<?hh

abstract class Base {
  abstract function foo(): Awaitable<int>;
}

class Alpha extends Base {
  async function foo(): Awaitable<int> {
    return 12;
  }
}

class Beta extends Base {
  async function foo(): Awaitable<int> {
    return 42;
  }
}

interface IBase {
  function bar(): Awaitable<bool>;
}

class One implements IBase {
  async function bar(): Awaitable<bool> {
    return true;
  }
}

class Two implements IBase {
  async function bar(): Awaitable<bool> {
    return false;
  }
}

async function test(Base $b, IBase $i): Awaitable<int> {
  await $b->foo();
  await $i->bar();
}
```

Reviewed By: jano, markw65

Differential Revision: D8056108

fbshipit-source-id: 1712ddbe0859657960b71404b75e5a6e378c6a0c