Serialize the call graph
Summary:
I didn't serialize ProfTransRec::mainCallers, or guardCallers, because
they're tc addresses, and I thought they were just used to smash the
callers of the profiled funcs after we optimized (plus, it doesn't
really make sense to record tc-addresses that won't exist when we
unserialize). But we need these to build the call graph.

This adds a set of TransID's representing the callers to each
Proflogue, for use when building the call graph.

Reviewed By: ottoni

Differential Revision: D8066398

fbshipit-source-id: a1cf48a9a7fb303f8e9280f81a8c5bf353c3856a