Introduce subtype simplification function
Summary:
As a first step to moving towards a constraint-based approach to subtyping, we add a function `simplify_subtype` that takes a subtype assertion t1 <: t2 and if possible "reduces" it to a set of simpler assertions that are equivalent. This is used for both the initial step of subtyping, and also for the constraint decomposition that is used when generating constraints from `instanceof` tests.

Some error messages change somewhat, mainly involving invariant and contravariant parameters, where the explanation "Considering that type argument is contravariant" (or invariant) is handled by wrapping types in new Reason constructors rather than doing a `try`.

There remains a lot of code duplication between `simplify_subtype` and the subtyping function itself. This will go away gradually with subsequent diffs.

Reviewed By: manzyuk

Differential Revision: D7935701

fbshipit-source-id: f66ce15d7a6b25aa2eaf32477014e962114235db