HHVM Debugger: Poll before writing on output thread
Summary:
Fixes a bug where we can get stuck trying to disconnect the client if the output thread is in a blocking write call to the socket, which has been uncleanly torn down. The write call can block until the socket timesout, which can be a long time.

This also fixes a bug where if a single call to write doesn't consume our entire buffer, we fail to retransmit what remains. Calls to write are supposed to check how many bytes were actually written.

Reviewed By: velocityboy

Differential Revision: D8034975

fbshipit-source-id: c6167a94f298c14c45c6cd49bdca193da454c6a7