Replace TypedValue* bases in member-operations.h with tv_lval
Summary:
Replace all uses of TypedValue* as a member base in
member-operations.h with tv_lval, and fix all related fallout. The JIT still
doesn't know anything about tv_val<>, which is fine as long as tv_val<>
contains nothing but a TypedValue* and is passed in a register.

MInstrState::base is a tv_lval now, rather than a union of tv_lval and
tv_rval. The latter would more accurately reflect what's going on, but the
former is consistent with how it was a TypedValue* before, with a few
const_casts as needed. The equivalent of const_cast<TypedValue*>() is
tv_rval::as_lval(), which is used in exactly the same places. It's probably
worth cleaning this up at some point, but let's stick to solving one problem at
a time.

Reviewed By: alexeyt

Differential Revision: D7997808

fbshipit-source-id: fd24ece55e3e81b9cbeea427729ae2d020bf8628