Fix a crash in ext_icu_ucsdet
Summary:
This has been broken since ext_icu was first written; an EncodingMatch
is only valid as long as its EncodingDetector hasn't been destroyed,
or re-used. This moves the ownership of the UCharsetDetector from the
EncodingDetector to the EncodingMatch.

At the same time, avoid converting from String->std::string in setText and
setDeclaredEncoding.

Reviewed By: paulbiss, edwinsmith

Differential Revision: D8076480

fbshipit-source-id: 29bad7e0e928d7dd11de2ce4ef4cbddd97b46871