Don't check for errors in decl mode parsing
Summary: In decl mode, we still check for parser errors. This takes up to **20%** of parsing time according to profiling. There's no need to check for errors when parsing for decl bodies: any file that could have a parsing error must have been edited directly, so we would need to full parse it anyways.

Reviewed By: vassilmladenov

Differential Revision: D8045109

fbshipit-source-id: cc9c4b79d7fc9b0d9bed2c79bb3ea0e4f35b90ff