Rewrite trivial (non-)equality comparison checks as a Tast_check
Summary:
Checks for trivial (non-)equality comparisons (e.g., comparing a provably nonnullable value to `null`, or comparing values of distinct primitive types known statically) are intertwined with type-checking of comparison expressions.  On the one hand, this clutters the type-checking code (and really, these checks don't belong into type-checking as the code they reject is safe wrt to the runtime but probably contains a logical error); on the other hand, they may be run too early before the types have been resolved sufficiently, resulting in some errors going unnoticed.

A better approach is to do these checks on a typed AST, similarly how sketchy null checks are performed (D7801280).  I chose to report errors for all comparisons that statically resolve to `true` or `false`, including those used inside conditions.  (Why do we not do this already?)

I plan to add instanceof checks that are statically known to be true or false to this check later, as well as checks for existence of shape fields that are statically know to exist (or to not exist).

Differential Revision: D8096020

fbshipit-source-id: 33c09a023a7db40b0979fd24b7029ff0a1b50578