make persistent rds handle not relative to tl_base
Summary:
The persistent region isn't request local anyways. So just use the same global address in all threads.
To distinguish between persistent handle and normal/local handles, strengthen the mode-based type system in rds::Link, along with some refactoring.

A persistent rds handle is interpreted as the offset to `s_persistent_base` - 4G.  When RDS_FIXED_PERSISTENT_BASE is defined (reasonable when `low_malloc()` guarantees the address returned is between 1G and 4G), `s_persistent_base` is set to 4G, and the handle is equal to the absolute address of the persistent data.

Reviewed By: markw65

Differential Revision: D7584137

fbshipit-source-id: 783039b08a0c863577e31db475173d61f9849a2a