Fix elaboration of HH autoimport class names when HH syntax is disabled
Summary:
In Hack files, class names which are autoimported from the HH namespace are elaborated into the HH namespace. For example, Container -> \HH\Container; Traversable -> \HH\Traversable. Other class names are elaborated into the current namespace.

In PHP files, we do not autoimport any classes from the HH namespace, and should not elaborate any names into that namespace. However, we currently treat any name which would be autoimported from \HH (in a Hack file) as a name in the global namespace, instead of elaborating it into the current namespace as we do with any other name.

After this change, we always elaborate names into the current namespace in PHP files (unless the EnableHipHopSyntax option is set), matching HHVM's behavior for the added test case.

fixes facebook/hhvm#8204

Differential Revision: D8100689