Convert is/as-expression hint validation to TAST check
Summary:
This moves the validation of hints in is/as-expressions out of Typing_utils to the TAST check phase, as a prerequisite to validating type aliases. It also cleans up some duplicated code in Typing.

Semantically, this implies a couple things:
- Because validation comes after refinement, we unconditionally refine types, even when they're invalid. However, because invalid types will be unfixmeable errors, it's a moot point.
- This prevents duplicate errors caused by expressions being typechecked more than once during the typing phase, since we're guaranteed that each node in the TAST is traversed once.

Both of the above changes are reflected in the regenerated tests.

Reviewed By: jamesjwu

Differential Revision: D8132622

fbshipit-source-id: 72c56d724700edac1086bd42c41ea407b72490db