Validate type aliases in is/as-expressions
Summary:
This adds the validation required to support type aliases in is/as-expressions. Because an invalid type may be hiding at any intermediary decl type on the way to localization, we must validate each of these decl types.

We do this by adding a validator on the ety_env. I did this because we already thread the ety_env through the recursive localization calls, which seemed cleaner than adding a new param to thread everywhere that `localize` was called.

Reviewed By: jamesjwu

Differential Revision: D8148890

fbshipit-source-id: cad781bf557e697153d1b7e8a2b26cd2b5debf33