Elide singleton unresolved's when setting local variables
Summary: When setting a local variable to a type we expand the type to prevent it from being affected by future unifications.  E.g., when setting a local to variable to `^(ty)`, we actually store a mapping to `(ty)`.  The unresolved wrapper here seems redundant: because of unbinding, `(ty)` cannot grow and hence //should// behave like `ty`.  This is not entirely accurate: e.g., unification of `(ty1)` and `ty2` currently only checks if `ty1` is a subtype of `ty2`; this, I think, is different from what unification would do to `ty1` and `ty2`.  Also, a subtype check `ty1 <: (ty2)` works by unifying `(ty1)` and `(ty2)`, which is in general different from the effect of a subtype check `ty1 <: ty2`.  However, I was unable to come up with example taking advantage of these discrepancies.

Differential Revision: D8146733

fbshipit-source-id: bd3f3ddca7879ac87559a7bbd0a3fcfe88e24c33