Pass this_ty as separate parameter in subtyping
Summary:
Previously we passed in a `this_ty` parameter associated with both subtype and supertype. But the latter did not get used, and does not really make sense.
So now we replace the "unification environment" with three parameters: `this_ty`, and `unwrappedToption_sub` and `unwrappedToption_super` that indicate whether we have "gone under" a `Toption` on the subtype or supertype.

Reviewed By: manzyuk

Differential Revision: D8135036

fbshipit-source-id: 61aa06e1466374a168fc0d5bc3b6bdec268caa4c