Simplify treatment of nullable in subtyping and unification
Summary:
It turns out that the logic associated with unwrapping nullables in subtyping and unification can be simplified substantially.
We only care about unwrapping of the supertype. For the case `?t1 <: ?t2`, we already reduce this to (iff) `t1 <: ?t2`. And the logic that acts on `unwrappedToption` for the supertype can be moved from unification into subtyping.

Reviewed By: dlreeves

Differential Revision: D8141037

fbshipit-source-id: 1b7f2fbb364a14262083b78926542247e584c6d1