hh check --single
Summary:
If you use Nuclide, we know what files you are looking at and we're going to be able to give you errors for them ASAP. But (unfortunately!) not everyone uses Nuclide, and I want to give them some error-reporting wins of non-blocking typechecking too.

This is basically `hh_single_type_check`, but with all the server state context available.

Note that (for simplicity) this is slightly different from we do in Nuclide - there we try to really always shows you all the errors “reported” in this file - it's possible that analysis of file B reports errors in file A, and we try to track this and make sure that we analyze B even if you don't have it open if you are looking at A. This command just returns errors uncovered when analyzing A. In most cases those two things should be the same - I'll rethink it if we get a lot of confused people (unlikely),

It also always reanalyzes the entire file, even if it's error list is already cached. It's hard to access error list while it is being actively recomputed, and for most files re-generating it's own error list is instantaneous.

(Note: this ignores all push blocking failures!)

Reviewed By: alexchow

Differential Revision: D8126276

fbshipit-source-id: 62462b3dee600409fab7887a20432f4a9740169e