Remove hh --ignore-fixme
Summary:
This does "typechecking on the side" by taking copy of original ServerEnv.env, updating it with some changes, running ServerTypeCheck.check with modified environment, and then ignoring the updated environment to return the original one.

This is very dangerous and I can't imagine is correct. There are tons of global state outside of environment that ServerTypeCheck.check updates, and the environment needs to be in sync with this global state. Because of this it's never safe to ignore the environment returned from ServerTypeCheck.check in incremental mode.

Added in {D4510524} as "auto-fix hackathon project" and mostly unused recently[1]. If someone needs it, it would need to be rewritten using something like ServerIdeUtils that uses local changes to better isolate global state from inadvertent modifications.

Reviewed By: oulgen

Differential Revision: D8132458

fbshipit-source-id: cf1323956291895020dd671eda23bc7416d2f640