Further restrict usage of the wildcard typehint
Summary: In our design discussion for the wildcard parameter, we decided that until we spec out true existential types, the wildcard should be heavily restricted. This restricts the wildcard such that even when it's allowed inside a particular expression, it may only be used in type parameter lists, and only in the first depth level. This means that of the hints `_`, `Foo<_>`, `Foo<Bar<_>>`, only the second is valid.

Reviewed By: manzyuk

Differential Revision: D7458816

fbshipit-source-id: b3ac47d1f632120186b51f04ad7270984e6ababa