Make sure rds::bind initializes the handle
Summary:
We recently found a race because TypeAliasReq was setting up the
handle and then initializing the value outside of bind, after the lock
had been released.

This ensures that bind initializes the handle. It only initializes it when
it's persistent, because this code path will only execute once in the
lifetime of the server, while a non-persistent handle needs to be
initialized on every request. ie code can't rely on bind to initialize a
non-persistent handle.

Reviewed By: binliu19

Differential Revision: D8166681

fbshipit-source-id: 5b817f4c2a44ca80bc1dd511382749b634b434a8