Add a widening rule for closed shapes
Summary:
Add a rule that captures the fact that optional fields can be added to a closed shape, so that, e.g., `shape('x' => int)` is considered a subtype of `shape('x' => int, ?'y' => string)`.

Also add a depth subtyping rule for open shapes (they are a separate syntactic category) and clean things up to have consistent format.

Reviewed By: CatherineGasnier

Differential Revision: D8161006

fbshipit-source-id: 0e3fb03aba67c692421cf9dbfe084b8b01a02add