Better error messages for is/as-expressions
Summary:
This solves a lack of polish on is/as-expression errors specifically for type aliases. If we encounter an invalid type during localization, the error would appear at the definition of the offending type alias, not at the usage of the is/as operator!

This adds an error at the hint as well as the type. In most cases, the positions will be the same, but when type aliases are involved, users will see an error in the correct place and a helpful pointer to the offending type alias.

Differential Revision: D8178649

fbshipit-source-id: 184496731dc233d289f1b1166141739b8f6f65e8