LSP supports unrecognized methods
Summary:
When the client invokes an unknown LSP method, what should the LSP server do? Well, in the case that it was a request, then the LSP spec mandates that the server send back an error response.

Prior to this diff we weren't doing this right. When we didn't know the method and therefore didn't know how to parse it into an ocaml data-structure, we threw an exception (without recording anything about what the kind of request was).

That's not enough. If we don't know the method was, we at least need to record the fact that it was an incoming request, and record its ID - enough information for us to send back a response.

Reviewed By: gabelevi

Differential Revision: D8175661

fbshipit-source-id: 6a3091028bf7123115a23bf78a9fa5a41ab714a1