Detect Shapes::keyExists() checks that statically resolve to true or false
Summary:
This is similar to trivial comparison checks.  A check `Shapes::keyExists($s, $k)`, where the key `$k` is statically known to exist in `$s` (e.g., because `$k` is declared as a required field in the type of `$s`) or where it is statically known to not exist in `$s` (e.g., because `$k` was unset before, or because the type of `$s` is a closed shape that doesn't allow `$k`), is reported as an error.  See the test cases for more examples.

The new behavior is hidden behind an experimental feature flag.

Reviewed By: CatherineGasnier

Differential Revision: D8122088

fbshipit-source-id: b4622f864bf12317d98ee0208b1f744f1822e674