Tast_check for invalid Shapes::idx() calls
Summary:
A check for when `Shapes::idx()` is used to access a field that is known not to exist (e.g., because it was unset previously, or because the first argument is known to be a closed shape that doesn't have the specified key declared).  See the test cases below for more examples.

Currently, this check only fires if the shape type is fully resolved.  Furthermore, it is hidden behind an experimental feature flag.

Reviewed By: CatherineGasnier

Differential Revision: D8163070

fbshipit-source-id: 83545cf71a3df4330f279558cefd00429d98ee43