Do not call onRequest if we have already sent a response.
Summary:
Previously the `ProxygenTransport` inside HHVM would happily call `onRequest`, which enqueues the request to execute the corresponding PHP endpoint, even if a response had already been sent (due to an error in initial handling of the request). This caused HHVM to do unnecessary work, and on debug instances is would trigger an assert when we try to respond to the request again.

To fix this we simply do not call `onRequest` if `m_sendEnded` is true, since there's nothing to be done with the request in this case.

This diff also makes the following changes:

- Fix up the bookkeeping in the `sendErrorResponse` case.
- Fix the access log listing any unknown HTTP method access as a GET.
- Change the proxygen-transport-tests to use `StrictMock`s and make the corresponding changes to the tests.

Reviewed By: afrind

Differential Revision: D8077886

fbshipit-source-id: c82d503c7d4a57066b5c35be80ee3490b91aedc7