Immutable foreach and catch bindings
Summary:
In hacksperimental, we allow immutable bindings for `foreach` and `catch` variables. This can be done by simply removing the `$` before the variable

For example:
```foreach ($arr as element) {
    // do something
}
```
```
foreach ($dict as k => v) {
    // do something
}
```
```
try {
    // throw exceptions
} catch (Exception e) {
    // catch it
}
```
Note that immutable binding does not guarantee const-ness. Method calls to an object reference can modify the object - it is only not possible to rebind the variable.

Reviewed By: arun11111

Differential Revision: D7845762

fbshipit-source-id: ec4298f4c2e937769567e109dd8b76ef58c36080