Add constructor type aliases
Summary: Typing for the PPL attribute will need to use the record field Env.inside_constructor, which we will set correctly in Tast_visitor. We use a type alias in the same style as static methods and variables in order to generate new visitor methods specifically for the constructor. We'll update Typing_env.env.inside_constructor for the visitor methods.

Differential Revision: D8208966

fbshipit-source-id: 431a4f88a6e32c218ba7885f47a639bc19108beb