Fix folding across blocks
Summary:
If a function that is simple enough to be foldable is passed a constant
argument coming from expression with control flow, hhbbc will propagate that
constant and fold the call. If a later call to the same function is not
foldable, it will mark the function as not foldable during each run of the
analyze phase, but this information will never reach the first call as each
analyze run starts from scratch.

However, the list of unfoldable functions is passed to the optimize phase,
which will prevent FPush from the first call from being inlined, but the FPass
and FCall calls will still be optimized away, as they are in a different block
that is processed independently and that block has an ActRec coming from the
analyze phase, which is marked as folded.

The above bug causes an invalid code to be generated. In the attached unit
test, hhbbc will change `FCallD "" "var_dump"` into `FCallD "" "foo"` to close
the mismatched FPushD, causing the wrong method to be called. Verifier catches
this in a debug mode as the outer FPI block is not properly closed, but if this
happens in a production mode, runtime would just execute the effects of foo()
twice (one folded, one real) and skip var_dump().

Let's fix this by revisiting all blocks with previously folded function calls
to the same method during analyze step to make analyze and optimize steps more
consistent.

Reviewed By: markw65

Differential Revision: D8201925

fbshipit-source-id: 6ac258702dcb91ad094f7409a8a167d56c24da91