Verify that shape fields in class property initializers contain only static code
Summary:
It seems that shapes were overlooked when this check was added in D2425912. This caused Hack to miss the function call in a shape initializer in D8212535.

After this change, we run this check on shape fields in class property initializers. Tests are also included for other array types.

Reviewed By: kmeht

Differential Revision: D8218788

fbshipit-source-id: 6d11fbeef93b2ea154b8dcc613bfaeccad518e9c