Use Typed AST in Linter_static_resources
Summary:
Convert Linter_static_resources from using typing hooks to examining a Typed AST.

In order to do this, this change adds Tast variants of the `static_string` and `ensure_static_string` helpers. Ideally, we would avoid code duplication by writing the implementation in a functor parameterized over AAST annotations (so that we can instantiate it once for the NAST and once for the TAST). Unfortunately, the `static_string_exn` function needs to be mutually recursive with the `get_const` function, which only operates on NASTs because only NASTs are stored in the typing heap. Duplicating `static_string_exn` (and all functions used to enter it) is the most straightforward way to work around this.

This change also adds an at_Call method to Tast_visitor.handler as a convenience.

Reviewed By: pittsw

Differential Revision: D8207514

fbshipit-source-id: 737f6eb1a0ce8e40a28b7bb0be8362f6dfcf17e5