Adjust LibXML ownership model to make reffiness unobservable
Summary:
The ownership model used by LibXML is unsafe to use with the hhbbc move optimization because tree ownership is held solely by the root node making it possible to observe the lifetime of the DOMElement or SimpleXMLElement holding the root of the tree.

In DOMDocument, PHP handles this by freeing the underlying LibXML xmlNodes out from underneath their associated DOMNodes and raising notices if these nodes are reused. With the move optimization it has become possible for properly structured code to trigger errors by freeing a seemingly unused root node, while child nodes are still live.

We use the property that each node may only belong to one tree, and that no node may descend from itself (e.g. no cycles) to implement a form of lazy refcounting on the tree. The way that DOMDocument and SimpleXML are structured it's not immediately obvious where we would need to inject reference counting to ensure that trees remain properly counted, however, we can lazily search for live descendants and construct reference counts when attempting to free tree roots.

Technically this diff represents a BC break in the way that LibXML is used in hhvm, and is therefore gated behind a new runtime option. The BC break itself should only be an issue if logic depends on the lifetime of the underlying tree or if the code was already malformed, both of which feel as though they are anti-patterns. The biggest risk with this diff is that we could transform a notice from accessing an invalid tree into a fatal from performing an invalid operation on a valid tree, but again the code was already broken in this case.

Reviewed By: markw65

Differential Revision: D8178043

fbshipit-source-id: 431a70110158b3064e8a66789ef9fb17cd73febc